# Speed up godef-jump by avoiding passing -t to godef