I found godef-jump taking more than 10 seconds to find the definition of a private function in a fairly small file (with lots of dependencies though).  This seemed suspicious so I dug into and found the -t flag made godef run much much slower:

% wc -l server.go
     127 server.go
% time godef -f server.go -t getFile > /dev/null
godef -f server.go -t getFile > /dev/null  11.99s user 15.19s system 221% cpu 12.271 total
% time godef -f server.go getFile > /dev/null
godef -f server.go getFile > /dev/null  0.77s user 0.95s system 205% cpu 0.834 total

godef-jump discards the type info anyway so with a bit of code we can avoid requesting it, making the call much faster (in my codebase anyway).